Added alot of things
Enemy Stats
Enemy Shooting
Enemy Taking damage
Enemy Doing Damage

+++ The Darkness Of The Depths/Assets/Scripts/WeaponStats/WeaponStats.cs
@@ -6,7 +6,7 @@ public abstract class WeaponStats : ScriptableObject
 {
     public string weaponName;
     public Sprite weaponSprite;
-    public float dmg;
+    public int dmg;
     public float knockBackForce;
     public float ArmorPenentration;
     public float range;
